# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Swagger Documentation of StockX server API",
    "version": "1.0.0",
    "title": "StockX API"
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "tags": [
    {
      "name": "User",
      "descriptions": "User endpoints"
    },
    {
      "name": "Product",
      "descriptions": "Products endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User data to add.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created user object.",
            "schema": {
              "$ref": "#/definitions/CreatedUser"
            }
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User data to login.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Logged user object.",
            "schema": {
              "$ref": "#/definitions/LoggedUser"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{uid}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Found user object.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product data to add.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created product object.",
            "schema": {
              "$ref": "#/definitions/CreatedProduct"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product data to update.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated product uid.",
            "schema": {
              "$ref": "#/definitions/UpdatedProduct"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Found list of products.",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{uid}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Found product object.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "user": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
            },
            "email": {
              "type": "string",
              "example": "lucas@email.com"
            },
            "password": {
              "type": "string",
              "example": "1234"
            },
            "name": {
              "type": "string",
              "example": "Lucas"
            },
            "created": {
              "type": "string",
              "example": "19/06/2023, 18:23:02"
            },
            "modified": {},
            "deleted": {}
          }
        }
      }
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "lucas@email.com"
        },
        "password": {
          "type": "string",
          "example": "1234"
        },
        "name": {
          "type": "string",
          "example": "Lucas"
        }
      },
      "required": [
        "email",
        "password",
        "name"
      ]
    },
    "CreatedUser": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "user": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
            },
            "email": {
              "type": "string",
              "example": "lucas@email.com"
            },
            "password": {
              "type": "string",
              "example": "1234"
            },
            "name": {
              "type": "string",
              "example": "Lucas"
            },
            "created": {
              "type": "string",
              "example": "19/06/2023, 18:23:02"
            },
            "modified": {},
            "deleted": {}
          }
        }
      }
    },
    "LoginUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "lucas@email.com"
        },
        "password": {
          "type": "string",
          "example": "1234"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "LoggedUser": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "user": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
            },
            "name": {
              "type": "string",
              "example": "Lucas"
            }
          }
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "product": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
            },
            "name": {
              "type": "string",
              "example": "Produto 1"
            },
            "amount": {
              "type": "number",
              "example": 1
            },
            "amountType": {
              "type": "string",
              "example": "Caixas"
            },
            "price": {
              "type": "number",
              "example": 10
            },
            "category": {
              "type": "string",
              "example": "Peças"
            },
            "created": {
              "type": "string",
              "example": "19/06/2023, 18:23:02"
            },
            "modified": {},
            "deleted": {}
          }
        }
      }
    },
    "Products": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string",
                "example": "SXn2duju1bS6uXRZBKgXoVbC43D3"
              },
              "name": {
                "type": "string",
                "example": "Produto 2"
              },
              "amount": {
                "type": "number",
                "example": 2
              },
              "amountType": {
                "type": "string",
                "example": "Caixas"
              },
              "price": {
                "type": "number",
                "example": 10
              },
              "category": {
                "type": "string",
                "example": "Peças"
              },
              "created": {
                "type": "string",
                "example": "19/06/2023, 18:25:02"
              },
              "modified": {},
              "deleted": {}
            }
          }
        }
      }
    },
    "AddProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Produto 1"
        },
        "amount": {
          "type": "number",
          "example": 1
        },
        "amountType": {
          "type": "string",
          "example": "Caixas"
        },
        "price": {
          "type": "number",
          "example": 10
        },
        "category": {
          "type": "string",
          "example": "Peças"
        }
      },
      "required": [
        "name",
        "amount",
        "amountType",
        "price",
        "category"
      ]
    },
    "CreatedProduct": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "product": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
            },
            "name": {
              "type": "string",
              "example": "Produto 1"
            },
            "amount": {
              "type": "number",
              "example": 1
            },
            "amountType": {
              "type": "string",
              "example": "Caixas"
            },
            "price": {
              "type": "number",
              "example": 10
            },
            "category": {
              "type": "string",
              "example": "Peças"
            },
            "created": {
              "type": "string",
              "example": "19/06/2023, 18:23:02"
            },
            "modified": {},
            "deleted": {}
          }
        }
      }
    },
    "UpdateProduct": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string",
          "example": "SXn2duju1bS6uXRZBKgXoVbC43D2"
        },
        "name": {
          "type": "string",
          "example": "Produto 1"
        },
        "amount": {
          "type": "number",
          "example": 1
        },
        "amountType": {
          "type": "string",
          "example": "Caixas"
        },
        "price": {
          "type": "number",
          "example": 10
        },
        "category": {
          "type": "string",
          "example": "Peças"
        }
      },
      "required": [
        "uid",
        "name",
        "amount",
        "amountType",
        "price",
        "category"
      ]
    },
    "UpdatedProduct": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "uid": {
          "type": "string",
          "example": "d17u0PhrETLAGNNHeONP"
        }
      }
    }
  }
}